refactor: break down large EncounterServiceImportExport into modular components
- Split 782-line service file into focused modules:
  * types.ts - Type definitions and schemas
  * dataBuilder.ts - Export data preparation
  * formatConverter.ts - XML/JSON conversion utilities
  * importProcessor.ts - Import data processing
  * sharingUtils.ts - Sharing and template functionality
- Reduced main service file to 178 lines (78% reduction)
- Created test helpers to reduce test duplication
- All 15 tests still pass with same coverage
- No breaking changes to public API

Resolves Codacy complexity and file size issues:
- File size reduced from 782 to 178 lines
- Test method complexity reduced by extracting helpers
- Improved maintainability and separation of concerns
